@@ -25,74 +25,75 @@
   "author": "Mark Judson",
   "license": "MIT",
   "dependencies": {
-    "@coreui/coreui-pro": "^2.1.8",
+    "@coreui/coreui-pro": "^2.1.14",
     "@coreui/icons": "^0.3.0",
-    "@fortawesome/fontawesome-pro": "^5.8.2",
-    "auth0-js": "^9.10.1",
+    "@fortawesome/fontawesome-pro": "^5.11.2",
+    "auth0-js": "^9.11.3",
     "autobind-decorator": "^2.4.0",
-    "babel-plugin-add-module-exports": "^1.0.0",
-    "express": "^4.16.4",
+    "babel-plugin-add-module-exports": "^1.0.2",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "fs": "0.0.1-security",
     "popper": "^1.0.1",
-    "react": "^16.8.2",
-    "react-bootstrap": "^1.0.0-beta.5",
+    "react": "^16.11.0",
+    "react-bootstrap": "^1.0.0-beta.14",
     "react-collapsy": "0.0.10",
-    "react-datepicker": "^2.5.0",
-    "react-dom": "^16.8.2",
+    "react-datepicker": "^2.9.6",
+    "react-dom": "^16.11.0",
     "react-light-accordion": "^0.1.4",
-    "react-redux": "^6.0.1",
+    "react-popper": "^1.3.4",
+    "react-redux": "^7.1.0",
     "react-redux-router": "0.0.14",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.9",
-    "react-select": "^2.4.3",
-    "react-spinners": "^0.5.4",
-    "react-toastify": "^5.1.0",
-    "redux": "^4.0.1",
-    "url-parse": "^1.4.4"
+    "react-router-dom": "^5.1.2",
+    "react-router-redux": "^4.0.8",
+    "react-select": "^3.0.8",
+    "react-spinners": "^0.6.1",
+    "react-toastify": "^5.4.0",
+    "redux": "^4.0.4",
+    "url-parse": "^1.4.7"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.3.3",
-    "@babel/node": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
-    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/cli": "^7.6.4",
+    "@babel/core": "^7.6.4",
+    "@babel/node": "^7.6.3",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-decorators": "^7.6.0",
     "@babel/plugin-transform-react-jsx": "^7.3.0",
-    "@babel/preset-env": "^7.3.1",
-    "@babel/preset-react": "^7.0.0",
-    "@coreui/coreui": "^2.1.7",
-    "babel-loader": "^8.0.5",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/preset-react": "^7.6.3",
+    "@coreui/coreui": "^2.1.12",
+    "babel-loader": "^8.0.6",
     "babel-plugin-react-transform": "^3.0.0",
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^4.3.1",
-    "cheerio": "^1.0.0-rc.2",
+    "cheerio": "^1.0.0-rc.3",
     "clean-webpack-plugin": "^1.0.1",
-    "color": "^3.1.0",
-    "concurrently": "^4.1.0",
-    "css-loader": "^2.1.0",
-    "es6-promise": "^4.2.6",
+    "color": "^3.1.2",
+    "concurrently": "^4.1.2",
+    "css-loader": "^2.1.1",
+    "es6-promise": "^4.2.8",
     "eventsource-polyfill": "^0.9.6",
     "file-loader": "^3.0.1",
-    "formik": "^1.5.1",
-    "fs": "0.0.1-security",
-    "history": "^4.7.2",
+    "formik": "^1.5.8",
+    "history": "^4.10.1",
     "html-webpack-plugin": "^3.2.0",
     "jquery": "^3.4.1",
     "keymirror": "^0.1.1",
-    "less": "^3.9.0",
+    "less": "^3.10.3",
     "less-loader": "^4.1.0",
     "npm-run-all": "^4.1.5",
-    "open": "0.0.5",
+    "open": "^7.0.0",
     "perfect-scrollbar": "^1.4.0",
     "prop-types": "^15.7.2",
     "redux-immutable-state-invariant": "^2.1.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^7.3.1",
     "style-loader": "^0.23.1",
-    "toastr": "^2.1.4",
-    "webpack": "^4.29.5",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-middleware": "^3.6.0",
-    "webpack-dev-server": "^3.2.0",
-    "webpack-hot-middleware": "^2.24.3"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-hot-middleware": "^2.25.0"
   }
 }
