@@ -3,14 +3,17 @@ import PropTypes from 'prop-types';
 import Form from 'react-bootstrap/Form';
 import Col from 'react-bootstrap/Col';
 import Select from 'react-select';
-import 'localcss/style.css';
 
 class SearchDropdown extends PureComponent {
   render() {
     const { name, label, size, onChange, placeholder, options, selectedOption, error } = this.props;
     let wrapperClass = '';
     if (error && error.length > 0) {
-      wrapperClass += " " + 'has-error';
+      wrapperClass += ' has-error';
+    }
+
+    const searchStyles = {
+      control: styles => ({ ...styles, height: '65%', backgroundColor: '#515b65', color: '#FFF' })
     }
 
     return (
@@ -22,17 +25,20 @@ class SearchDropdown extends PureComponent {
           onChange={onChange}
           options={options}
           placeholder={placeholder}
-          className='select-text-color'
           isClearable={true}
           theme={(theme) => ({
             ...theme,
             borderRadius: 5,
+            borderColor: 'transparent',
             colors: {
-              ...theme.colors
+              ...theme.colors,
+              neutral80: 'white',  // selection color
+              neutral0: '#515b65', // select box background
+              primary25: 'green'   // list hoover
             }
           })}
         />
-        {error && <div className='alert alert-danger'>{error}</div>}
+        {error && <div className="alert alert-danger">{error}</div>}
       </Col>
     );
   }
